Rename godot::engine -> godot::classes
Keep all symbols available as deprecated.
