+++ godot-core/src/classes/manual_extensions.rs
